requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
aiohttp==3.9.1
python-dotenv==1.0.0
pillow==10.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
redis==5.0.1
